+++ pygo-tools/setup.py
@@ -2,7 +2,7 @@ from setuptools import find_packages, setup
 
 
 setup(
-    name='setup-util',
+    name='pygo-tools',
     version='0.1.0',
     packages=find_packages(),
     include_package_data=True,
